mental_health_backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.9.0
firebase-admin==7.1.0
python-dotenv==1.1.1
celery==5.5.3
twilio==9.8.2
pandas==2.3.2
openpyxl==3.1.5
# Machine Learning for Intent Recognition
scikit-learn>=1.3.0
nltk>=3.8.1
scipy>=1.11.0






